refactor: 입출력 클래스 변경  - Manager -> View로 클래스 네임 변경  - model -> view 디렉토리 변경
