+++ src/main/java/view/InputView.java
@@ -1,8 +1,8 @@
-package model.system;
+package view;
 
 import camp.nextstep.edu.missionutils.Console;
 
-public class InputManager {
+public class InputView {
 
     static String ReadInput() {
         return Console.readLine();
