+++ src/main/java/view/OutputView.java
@@ -2,7 +2,7 @@ package view;
 
 import model.referee.GameScoreboard;
 
-public class OutputManager {
+public class OutputView {
 
     public static void Write(GameScoreboard scoreboard) {
